pkb/requirements.txt
psycopg2-binary==2.9.9
minio==7.1.15
python-dotenv==1.0.0
alembic==1.12.0
fastapi==0.103.1
uvicorn==0.23.2
